src/requirements.txt
pandas
numpy
scipy
streamlit
autogluon
pymongo
python-dotenv
ipython
autoviz
joblib
tensorflow
keras-tuner
prophet
tslearn
optuna
shap
scikit-learn
statsmodels
xgboost
tf-keras
streamlit-oauth
google-cloud-bigquery
db-dtypes
google-cloud-bigquery-storage
requests
PyJWT
extra_streamlit_components

